+++ js/photo-data/photo-сonstants.js
@@ -1,0 +1,54 @@
+export const COUNTER = 25;
+export const MAXNUMBEROFCOMMENTS = 30;
+export const NUMBEROFAVATARS = 6;
+
+export const PHOTODESCRIPTION = [
+  'Закат на берегу океана с яркими оранжевыми оттенками.',
+  'Высокий горный пик, скрытый в облаках.',
+  'Уличный кот, отдыхающий на старой деревянной лавочке.',
+  'Чашка горячего кофе на столе у окна в дождливый день.',
+  'Пышный зелёный лес после летнего дождя.',
+  'Пара на фоне ночного мегаполиса с огнями города.',
+  'Стадо овец пасется на холмистой местности.',
+  'Радуга над полем с подсолнухами после ливня.',
+  'Уютная библиотека с полками, заполненными книгами.',
+  'Ручей, протекающий через лесную поляну.',
+  'Лавочка в парке под ярко-желтыми осенними деревьями.',
+  'Заброшенная ферма с ржавым трактором на фоне.',
+  'Маленький домик у озера с горами вдали.',
+  'Яркие воздушные шары, плывущие по голубому небу.',
+  'Туманное утро на озере с плавающими утками.',
+  'Закат над горами с розовым и фиолетовым небом.',
+  'Пляж с белым песком и кристально чистой водой.',
+  'Старая железнодорожная станция в полупустынной местности.',
+  'Яхты, пришвартованные в тихой морской гавани.',
+  'Утро в саду с яркими цветами и бабочками.',
+  'Рыбаки на деревянной лодке посреди озера.',
+  'Звездное небо над спокойной деревней ночью.',
+  'Тропинка в лесу, освещенная солнечными лучами.',
+  'Каменный мост через бурлящую реку.',
+  'Чистое ночное небо с полной луной над полем.'
+];
+
+export const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+export const NAMES = [
+  'Пафнутий', 'Ермолай', 'Тарас Барас', 'Фока', 'Семён Семёныч',
+  'Дормидонт', 'Елизар', 'Ферапонт', 'Суслопар', 'Прокоп',
+  'Тимоша Пузиков', 'Трофим Шалун', 'Кузьма Ляпунов', 'Никодим Жмурин', 'Зиновий Храпун',
+  'Яков Лупен', 'Панкрат Сурков', 'Федот Бубен', 'Аркаша Муляжов', 'Клим Вислогуб',
+  'Тихон Утюгов', 'Афанасий Лопухов', 'Яша Борщев', 'Гришка Окаянный', 'Игнат Косоротов',
+  'Матвей Плеткин', 'Устин Угольников', 'Порфирий Кашевар', 'Остап Сморчок', 'Никифор Чугунов'
+];
+
+export const Likes = {
+  MIN: 15,
+  MAX: 100,
+};
